+++ packages/select/src/ListMenu.tsx
@@ -24,10 +24,6 @@ interface ListMenuProps {
   children: React.ReactNode;
   id: string;
   referenceElement: React.MutableRefObject<HTMLElement | null>;
-  onClose: () => void;
-  onSelectFocusedOption: React.KeyboardEventHandler;
-  onFocusPreviousOption: () => void;
-  onFocusNextOption: () => void;
   className?: string;
   usePortal?: boolean;
   labelId?: string;
@@ -39,10 +35,6 @@ const ListMenu = React.forwardRef<HTMLUListElement, ListMenuProps>(
       children,
       id,
       referenceElement,
-      onClose,
-      onFocusPreviousOption,
-      onFocusNextOption,
-      onSelectFocusedOption,
       className,
       usePortal,
       labelId,
@@ -55,49 +47,6 @@ const ListMenu = React.forwardRef<HTMLUListElement, ListMenuProps>(
     const sizeSet = sizeSets[size];
 
     const ref = useForwardedRef(forwardedRef, null);
-
-    const onKeyDown = useCallback(
-      (event: React.KeyboardEvent) => {
-        // No support for modifiers yet
-        /* istanbul ignore if */
-        if (event.ctrlKey || event.shiftKey || event.altKey) {
-          return;
-        }
-
-        let bubble = false;
-
-        switch (event.keyCode) {
-          case keyMap.Tab:
-          case keyMap.Enter:
-            onSelectFocusedOption(event);
-            break;
-          case keyMap.Escape:
-            onClose();
-            break;
-          case keyMap.ArrowUp:
-            onFocusPreviousOption();
-            break;
-          case keyMap.ArrowDown:
-            onFocusNextOption();
-            break;
-          /* istanbul ignore next */
-          default:
-            bubble = true;
-        }
-
-        /* istanbul ignore else */
-        if (!bubble) {
-          event.preventDefault();
-          event.stopPropagation();
-        }
-      },
-      [
-        onClose,
-        onFocusNextOption,
-        onFocusPreviousOption,
-        onSelectFocusedOption,
-      ],
-    );
 
     const viewportSize = useViewportSize();
 
@@ -127,12 +76,13 @@ const ListMenu = React.forwardRef<HTMLUListElement, ListMenuProps>(
         refEl={referenceElement}
         usePortal={usePortal}
       >
+        {/* Keyboard events handled in Select component through event listener hook */}
+        {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events */}
         <ul
           aria-labelledby={labelId}
           role="listbox"
           ref={ref}
           tabIndex={-1}
-          onKeyDown={onKeyDown}
           onClick={onClick}
           className={cx(
             menuStyle,
