+++ packages/select/src/Select.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
 import { css, cx } from '@leafygreen-ui/emotion';
-import { useViewportSize } from '@leafygreen-ui/hooks';
-import { IdAllocator, OneOf } from '@leafygreen-ui/lib';
+import { useViewportSize, useEventListener } from '@leafygreen-ui/hooks';
+import { IdAllocator, OneOf, keyMap } from '@leafygreen-ui/lib';
 import { fontFamilies, breakpoints } from '@leafygreen-ui/tokens';
 import { colorSets, mobileSizeSet, Mode, Size, sizeSets } from './styleSets';
 import ListMenu from './ListMenu';
@@ -52,7 +52,7 @@ export type Props = {
     } & {
       onChange?: (
         value: string,
-        event: React.MouseEvent | React.KeyboardEvent,
+        event: React.MouseEvent | KeyboardEvent,
       ) => void;
       readOnly?: false;
     })
@@ -61,7 +61,7 @@ export type Props = {
       | {
           onChange: (
             value: string,
-            event: React.MouseEvent | React.KeyboardEvent,
+            event: React.MouseEvent | KeyboardEvent,
           ) => void;
           readOnly?: false;
         }
@@ -213,10 +213,7 @@ export default function Select({
   }, [children, uncontrolledSelectedOption, value]);
 
   const onSelect = useCallback(
-    (
-      option: OptionElement | null,
-      event: React.MouseEvent | React.KeyboardEvent,
-    ) => {
+    (option: OptionElement | null, event: React.MouseEvent | KeyboardEvent) => {
       if (value === undefined) {
         setUncontrolledSelectedOption(option);
       }
@@ -243,7 +240,7 @@ export default function Select({
   );
 
   const onDeselect = useCallback(
-    (event: React.KeyboardEvent) => {
+    (event: KeyboardEvent) => {
       onSelect(null, event);
     },
     [onSelect],
@@ -274,7 +271,7 @@ export default function Select({
   }, [children, allowDeselect]);
 
   const onSelectFocusedOption = useCallback(
-    (event: React.KeyboardEvent) => {
+    (event: KeyboardEvent) => {
       if (focusedOption !== undefined) {
         onSelect(focusedOption, event);
       }
@@ -283,12 +280,8 @@ export default function Select({
   );
 
   const onFocusFirstOption = useCallback(() => {
-    if (allowDeselect) {
-      setFocusedOption(null);
-    } else {
-      setFocusedOption(enabledOptions[0]);
-    }
-  }, [allowDeselect, enabledOptions]);
+    setFocusedOption(enabledOptions[0]);
+  }, [enabledOptions]);
 
   const onFocusLastOption = useCallback(() => {
     setFocusedOption(enabledOptions[enabledOptions.length - 1]);
@@ -333,6 +326,49 @@ export default function Select({
     [disabled],
   );
 
+  const onKeyDown = useCallback(
+    (event: KeyboardEvent) => {
+      // No support for modifiers yet
+      if (event.ctrlKey || event.shiftKey || event.altKey) {
+        return;
+      }
+
+      switch (event.keyCode) {
+        case keyMap.Escape:
+          onClose();
+          break;
+        case keyMap.Enter:
+          onSelectFocusedOption(event);
+          break;
+        case keyMap.ArrowUp:
+          if (!open && document.activeElement === menuButtonRef.current) {
+            onOpen();
+          }
+
+          onFocusPreviousOption();
+          break;
+        case keyMap.ArrowDown:
+          if (!open && document.activeElement === menuButtonRef.current) {
+            onOpen();
+          }
+
+          onFocusNextOption();
+          break;
+      }
+    },
+    [
+      onClose,
+      onOpen,
+      open,
+      menuButtonRef,
+      onFocusNextOption,
+      onFocusPreviousOption,
+      onSelectFocusedOption,
+    ],
+  );
+
+  useEventListener('keydown', onKeyDown);
+
   /**
    * Rendering
    */
@@ -361,10 +397,6 @@ export default function Select({
   const deselectionOption = useMemo(() => {
     const selected = selectedOption === null;
 
-    if (!allowDeselect) {
-      return null;
-    }
-
     return (
       <InternalOption
         className={undefined}
@@ -388,7 +420,6 @@ export default function Select({
     getOptionFocusHandler,
     placeholder,
     selectedOption,
-    allowDeselect,
   ]);
 
   const renderedChildren = useMemo(
@@ -501,11 +532,8 @@ export default function Select({
               : placeholder
           }
           deselected={selectedOption === null}
-          onDeselect={onDeselect}
           onOpen={onOpen}
           onClose={onClose}
-          onFocusFirstOption={onFocusFirstOption}
-          onFocusLastOption={onFocusLastOption}
           aria-labelledby={labelId}
           aria-controls={menuId}
           aria-expanded={open}
@@ -517,16 +545,12 @@ export default function Select({
             id={menuId}
             referenceElement={menuButtonRef}
             ref={listMenuRef}
-            onClose={onClose}
-            onSelectFocusedOption={onSelectFocusedOption}
-            onFocusPreviousOption={onFocusPreviousOption}
-            onFocusNextOption={onFocusNextOption}
             usePortal={usePortal}
             className={css`
               width: ${menuButtonRef.current?.clientWidth}px;
             `}
           >
-            {deselectionOption}
+            {allowDeselect && deselectionOption}
             {renderedChildren}
           </ListMenu>
         </MenuButton>
